Move EstatusLiquidacion enum to CORE
